Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/path-objects/2d.path.stroke.scale2-expected.txt

2d.path.stroke.scale2
Stroke line widths are scaled by the current transformation matrix
Actual output:
Expected output:


FAIL Stroke line widths are scaled by the current transformation matrix assert_equals: Red channel of the pixel at (0, 0) expected 0 but got 255
